requirements.txt
boto3==1.26.9
requests==2.28.1
pysnyk==0.9.3
